Add filter package (#513)
This package will be useful for a simple definition of how filters work, as well as a way to process tickets without relying on indexes. This will allow other filter types (eg, strings) to be added without a redis implementation. See package documentation for some more details.

I plan on replacing the current redis indexing with a "all tickets" index to remove the edge cases it gets wrong that we don't want to spend time fixing for v0.6. Instead this package will be responsible for filtering which tickets to return. This also removes the index configuration problem from v0.6. Then for v0.7, once the indexing and database solutions are chosen, we can go back to implementing the correct way to be indexing tickets.

Also included are test cases, separated in their own definition. These test cases should be used in the future for end to end tests, and for tests on indexes. This will help ensure that the system as a whole maintains the behavior specified here.