Add image for demo (#511)
This is barebones work to get an image for the demo working. This image will eventually contain the demo go-routines that emulate the clients, and presenting a dashboard for the state of the demo. Will followup with actual logic in the demo itself.

TESTED=Manually started om and the demo, ran proxy-demo and got the expected 404.