Autogenerate image commands based on a single list (#714)
With this change, anything added to the cmd/ folder is automatically
made into an image and included the image commands. Additional
images are also included. This does remove some commands that are
for pushing specific sets of images, but it seems rare/never (asking yfei1
and sawagh) that anyone is actually using these commands.

Includes some commenting to hopefully alleviate the magic being added.