Store indexes used for a ticket, and use them to deindex (#789)
This has two primary advantages:
The redis key of the index can be decided at ticket index / query time. This is important for string equal indexing, where the plan is to concatenate the OM index name and the string value to form the redis key.
Improved correctness when indexes are changed: The ticket will now clean up the indexes it was created with, preventing old indices from existing after all the tickets that used them are gone.

This does add an extra read when deindexing a ticket, but I think the correctness improvement alone is worth that.

Other notes:
Turns out the indexes need to be in a list of interface{} to concatenate with the redis key for the cache, so I changed my mind about computing the list in extractIndexedFields. So extractIndexedFields instead just returns the map of index to values.

Improved a test's assertions by using ElementsMatch.