Implement basic tunable parameters logic for benchmarking scenarios (#1030)
This commit implements the knobs to control ShouldCreateTicketForever, ShouldAssignTicket, ShouldDeleteTicket, TicketCreatedQPS, and CreateTicketNumber. Also removed the roster-based-mmf from the repo since it is only used for the scale test and there is no need to build its image in every CI run.

After this commit got checked in, users are able to configure the knobs via the new benchmarking framework and run make install-scale-chart to it.

TODO:

Implement the filter number and profile number logic. This requires a rewrite for examples/scale/tickets and examples/scale/profiles package.