Changed Synchronizer proto to return matchIDs instead (#1080)
This commit changed Synchronizer proto to return matchIDs instead. Also bumped up the numbers of the unnamed channels in synchronizer starting from m3c and changed the channel type starting from m4c to chan string as the next step of the API change is to have the evaluator returns the match ids instead.