Add omerrors and use it it in backend_service and evaluator_client (#1081)
Two methods are added:

ProtoFromErr: returns a grpc status given an error, with some reasoned handling for special cases. This will be used to set errors onto the FetchMatchesSummary in a followup PR.
-WaitOnErrors: this allows some number of functions to run that will all return errors. The first to return an error will be the error returned overall, and it ensures all go routines finish.
WaitOnErrors is used to simplify code in backend_service and the grpc portion of evaluator_client.

Also I realized that synchronizeSend should specify which context is being used where better.