Fix install-scale-chart target (#1322)
* Implement backfill querying

* Update location for stable and incubator charts

* Add MMF backfill example

* Simplify MMF backfill example

* Render jaeger configuration if it is enabled

Helm fails to install open-match chart with disabled jaeger because it cannot find
openmatch.jaeger.agent template which is declared in jaeger subchart. Helm is not able to
find that template because jaeger subchart is not loaded because it is marked as disabled
in open-match chart dependencies.

* Update install-scale-chart target

Currently open-match-scale subchart is installed separately from open-match chart but they are tightly coupled.
Pods declared in scale subchart have dependencies on service accounts, config maps provisioned by open-match
chart. So the problem is that helm renders incorrect service account, config map names. It can be fixed by
specifying explicit names in install-scale-chart target.

Co-authored-by: Alexander Apalikov <alexander.apalikov@globant.com>